refactoring : split main.cpp into main_ui.cpp
