+++ src/main_ui.hpp
@@ -1,0 +1,11 @@
+// Copyright 2022 Arthur Sonzogni. All rights reserved.
+// Use of this source code is governed by the MIT license that can be found in
+// the LICENSE file.
+#ifndef JSON_TUI_MAIN_UI_HPP
+#define JSON_TUI_MAIN_UI_HPP
+
+#include <nlohmann/json.hpp>
+
+void DisplayMainUI(const nlohmann::json& json, bool fullscreen);
+
+#endif /* json_tui_main_ui_hpp */
