+++ src/mytoggle.cpp
@@ -10,12 +10,8 @@ using namespace ftxui;
 
 class MyToggleImpl : public ComponentBase {
  public:
-  MyToggleImpl(const char* label_on,
-               const char* label_off,
-               bool* state)
-      : label_on_(label_on),
-        label_off_(label_off),
-        state_(state) {}
+  MyToggleImpl(const char* label_on, const char* label_off, bool* state)
+      : label_on_(label_on), label_off_(label_off), state_(state) {}
 
  private:
   // Component implementation.
