+++ src/version.hpp.in
@@ -1,1 +1,9 @@
+// Copyright 2022 Arthur Sonzogni. All rights reserved.
+// Use of this source code is governed by the MIT license that can be found in
+// the LICENSE file.
+#ifndef JSON_TUI_VERSION_HPP
+#define JSON_TUI_VERSION_HPP
+
 static const char project_version[] = "@CMAKE_PROJECT_VERSION@";
+
+#endif  // JSON_TUI_VERSION_HPP
